refactor: enhance device information output and improve code organization
