+++ src/quri_parts_oqtopus/backend/estimation.py
@@ -10,9 +10,9 @@ from quri_parts.core.operator import Operator
 from quri_parts.openqasm.circuit import convert_to_qasm_str
 
 from quri_parts_oqtopus.backend.config import (
-    DateTimeEncoder,
     OqtopusConfig,
 )
+from quri_parts_oqtopus.backend.utils import DateTimeEncoder
 from quri_parts_oqtopus.rest import (
     ApiClient,
     Configuration,
